Repairs after rebase.
